SimBridge now templated to publish all deformable meshes in simulation, regardless of simulation type. Added launch argument for simulation type to the launch files
