+++ ros_workspace/src/sim_bridge/src/sim_bridge_node.cpp
@@ -1,6 +1,7 @@
 #include <cstdio>
 
 #include "sim_bridge/SimBridge.hpp"
+#include "sim_bridge/VirtuosoSimBridge.hpp"
 
 #include "config/simulation/VirtuosoTissueGraspingSimulationConfig.hpp"
 #include "simulation/VirtuosoTissueGraspingSimulation.hpp"
@@ -30,29 +31,11 @@ void runSim(Sim::Simulation* sim)
     sim->run();
 }
 
-int main(int argc, char ** argv)
+template<typename SimulationType>
+void startNode(SimulationType* sim)
 {
-    rclcpp::init(argc, argv);
-
-
-    // parse command line arguments for config filename
-    std::string config_filename;
-    for (int i = 1; i < argc; i++)
-    {
-        std::string arg = argv[i];
-        if (arg == "--config-filename" && i+1 < argc)
-        {
-            config_filename = argv[++i];
-        }
-    }
-
-    // create the simulation config object from the yaml config file
-    Config::VirtuosoTissueGraspingSimulationConfig config(YAML::LoadFile(config_filename));
-    // create the simulation from the config object
-    Sim::VirtuosoTissueGraspingSimulation sim(&config);
-
     // start up the simulation in a separate thread
-    std::thread sim_thread(runSim, &sim);
+    std::thread sim_thread(runSim, sim);
 
     // wait for the simulation to be set up
     // required for loading meshes, setting up sim objects, etc.
@@ -62,10 +45,62 @@ int main(int argc, char ** argv)
     }
 
     // then start up the SimBridge ROS node
-    rclcpp::spin(std::make_shared<SimBridge>(&sim));
+    rclcpp::spin(std::make_shared<SimBridge<SimulationType>>(sim));
 
     sim_thread.join();
 
     rclcpp::shutdown();
+}
+
+int main(int argc, char ** argv)
+{
+    rclcpp::init(argc, argv);
+
+
+    // parse command line arguments for config filename and simulation type
+    std::string config_filename;
+    std::string simulation_type;
+    for (int i = 1; i < argc; i++)
+    {
+        std::string arg = argv[i];
+        if (arg == "--config-filename" && i+1 < argc)
+        {
+            config_filename = argv[++i];
+        }
+
+        if (arg == "--simulation-type" && i+1 < argc)
+        {
+            simulation_type = argv[++i];
+        }
+    }
+
+    if (simulation_type == "VirtuosoTissueGraspingSimulation")
+    {
+        // create the simulation config object from the yaml config file
+        Config::VirtuosoTissueGraspingSimulationConfig config(YAML::LoadFile(config_filename));
+        // create the simulation from the config object
+        Sim::VirtuosoTissueGraspingSimulation sim(&config);
+
+        startNode<Sim::VirtuosoSimulation>(&sim);
+    }
+    else if (simulation_type == "VirtuosoSimulation")
+    {
+        Config::VirtuosoSimulationConfig config(YAML::LoadFile(config_filename));
+        Sim::VirtuosoSimulation sim(&config);
+
+        startNode<Sim::VirtuosoSimulation>(&sim);
+    }
+    else if (simulation_type == "Simulation")
+    {
+        Config::SimulationConfig config(YAML::LoadFile(config_filename));
+        Sim::Simulation sim(&config);
+        startNode<Sim::Simulation>(&sim);
+    }
+    else
+    {
+        std::cerr << "Unrecognized simulation type: " << simulation_type << std::endl;
+        assert(0);
+    }
+
     return 0;
 }
